+++ truffle/src/com.oracle.truffle.dsl.processor/src/com/oracle/truffle/dsl/processor/TruffleSuppressedWarnings.java
@@ -69,8 +69,9 @@ public final class TruffleSuppressedWarnings {
     public static final String ASSUMPTION = "truffle-assumption";
     public static final String GUARD = "truffle-guard";
     public static final String DEPRECATION = "deprecation";
+    public static final String INTERPRETED_PERFORMANCE = "truffle-interpreted-performance";
     public static final List<String> ALL_KEYS = List.of(ALL, TRUFFLE, STATIC_METHOD, LIMIT, UNUSED, NEVERDEFAULT, INLINING_RECOMMENDATION, SHARING_RECOMMENDATION, ABSTRACT_LIBRARY_EXPORT,
-                    DEPRECATION);
+                    DEPRECATION, INTERPRETED_PERFORMANCE);
 
     public static Set<String> getWarnings(Element element) {
         AnnotationMirror currentWarnings = ElementUtils.findAnnotationMirror(element, SuppressWarnings.class);